changed Tool class, modified accordingly Brush, Bucket; added Knife tool

+++ arthropod_describer/common/photo.py
@@ -66,7 +66,7 @@ class LabelImg:
     def reload(self):
         # TODO remove the `1000 * ` part
         img = 1000 * np.logical_not(io.imread(str(self._path)))
-        self.label_img = img
+        self.label_img = img.astype(np.uint16)
         self.color_map = {label: QColor.fromRgb(*tuple(np.random.randint(0, 256, (3,))), 100).rgba() for label in np.unique(self.label_img)}
         self.color_map[0] = QColor.fromRgb(0, 0, 0, 0).rgba()
